requirements.txt
faiss-cpu
numpy
psycopg2-binary
sentence-transformers
python-dotenv
python-docx
langchain-core
langchain-groq
groq

